Move EntityDeleter
